Add images for major update
All the images required for the new version of the site to run. This includes: a replacement background, with both a dark mode and light mode version; a picture of Chadsworth the Hippo (serves as a logo and favicon); a picture of Emma Mei-Li; dark mode and light mode versions of the dark mode/light mode toggle switch; and dark mode and light mode versions of the social media icons that correspond to Emma's primary social media. See the patch notes from today's major update for more information.